+++ cc-service-api/src/main/java/com/jtmc/apps/reforma/api/v1/persona/PersonaApi.java
@@ -12,9 +12,20 @@ public interface PersonaApi {
     @Produces(MediaType.APPLICATION_JSON)
     PersonaResponseList selectAll();
 
+    @GET
+    @Path("/{personaId}")
+    @Produces(MediaType.APPLICATION_JSON)
+    PersonaResponse selectOne(@PathParam("personaId") int personaId);
+
     @POST
     @Path("/")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     Response insert(PersonaRequest request);
+
+    @POST
+    @Path("/{personaId}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    Response update(@PathParam("personaId") int personaId, PersonaRequest request);
 }