+++ CHANGELOG.md
@@ -7,6 +7,7 @@ release date: -
 - change initial-amount POST action to it's own api endpoint
 - add timezoneId and status to branch request
   - this will allow basic branch/account setup
+- Persona API to allow CRUD operations
 
 ## 2.3.2
 release date: May 30, 2024